[Stack Functional Integration] Add minimal hearbeat tests  (#112986)
* add hearbeat tests and convert to TS

* fix ts errors
