+++ x-pack/test/stack_functional_integration/apps/filebeat/index.ts
@@ -4,10 +4,10 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+import { FtrProviderContext } from '../../../functional/ftr_provider_context';
 
-export default function ({ loadTestFile }) {
-  describe('heartbeat app', function () {
-    require('./_heartbeat');
-    loadTestFile(require.resolve('./_heartbeat'));
+export default function ({ loadTestFile }: FtrProviderContext) {
+  describe('filebeat app', function () {
+    loadTestFile(require.resolve('./filebeat'));
   });
 }
