+++ x-pack/test/stack_functional_integration/apps/metricbeat/index.ts
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+import { FtrProviderContext } from '../../../functional/ftr_provider_context';
+
+export default function ({ loadTestFile }: FtrProviderContext) {
+  describe('metricbeat app', function () {
+    loadTestFile(require.resolve('./_metricbeat'));
+    loadTestFile(require.resolve('./_metricbeat_dashboard'));
+  });
+}
