+++ x-pack/test/stack_functional_integration/apps/packetbeat/index.ts
@@ -4,9 +4,10 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+import { FtrProviderContext } from '../../../functional/ftr_provider_context';
 
-export default function ({ loadTestFile }) {
-  describe('filebeat app', function () {
-    loadTestFile(require.resolve('./filebeat'));
+export default function ({ loadTestFile }: FtrProviderContext) {
+  describe('packetbeat app', function () {
+    loadTestFile(require.resolve('./_packetbeat'));
   });
 }
