--- x-pack/test/stack_functional_integration/apps/winlogbeat/index.js
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export default function ({ loadTestFile }) {
-  describe('winlogbeat app', function () {
-    loadTestFile(require.resolve('./_winlogbeat'));
-  });
-}
